Release 0.4.5

See merge request elympics/unity-sdk-package!57

+++ Runtime/InputsAndSnapshots/Snapshots/ElympicsVars/Comparers/ElympicsIntEqualityComparer.cs
@@ -9,6 +9,6 @@ namespace Elympics
 		{
 		}
 
-		protected override float Distance(int a, int b) => Math.Abs(a - b);
+		protected override float Distance(int a, int b) => Math.Abs((long)a - b);
 	}
 }
